Adjust project structure and incorporate new project naming conventions
