--- dashboard/app/Pages/Index.cshtml
@@ -1,4 +1,0 @@
-﻿@page
-@model IndexModel
-
-<p>@Model.Message</p>
