--- dashboard/app/Pages/Shared/_Layout.cshtml
@@ -1,10 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <title>Dashboard</title>
-</head>
-<body>
-    @RenderBody()
-</body>
-</html>
