--- dashboard/app/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using Dashboard
-@namespace Dashboard.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
