--- dashboard/app/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-﻿{
-  "profiles": {
-    "Dashboard": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "applicationUrl": "https://localhost:5001;http://localhost:5000",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
