+++ dashboard/src/Piipan.Dashboard/Pages/Index.cshtml
@@ -1,0 +1,4 @@
+﻿@page
+@model IndexModel
+
+<p>@Model.Message</p>
