+++ dashboard/src/Piipan.Dashboard/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using Piipan.Dashboard
+@namespace Piipan.Dashboard.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
