+++ dashboard/src/Piipan.Dashboard/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+﻿{
+  "profiles": {
+    "Piipan.Dashboard": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "applicationUrl": "https://localhost:5001;http://localhost:5000",
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      }
+    }
+  }
+}
